requirements.txt
requests>=2.13.0
json-encoder>=0.4.4
python-status>=1.0.1
aiohttp>=2.2.0
